pyproject.toml
[tool.poetry]
name = "k1core"
version = "0.1.0"
description = ""
authors = ["Dmytro Podolynnyi"]
readme = "README.md"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.4"
fastapi = "^0.115.8"
uvicorn = {extras = ["standard"], version = "^0.17.6"}
pydantic = "^1.9.1"
celery = "^5.4.0"
flower = "^2.0.1"
redis = "^5.2.1"
requests = "^2.32.3"
django-environ = "^0.8.1"
psycopg2 = "^2.9.10"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.20"
